Extract test data creation into data test data factory
