+++ package.json
@@ -44,7 +44,7 @@
     "transform": {
       ".(ts|tsx)": "ts-jest"
     },
-    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
+    "testRegex": "\\.(test|spec)\\.(ts|tsx|js)$",
     "moduleFileExtensions": [
       "ts",
       "tsx",
